moved MenuRow to a diff folder to prevent build issues
